Rename rootReducer.js to index.js and fix imports
