+++ src/store.js
@@ -1,6 +1,6 @@
 import { createStore, applyMiddleware, compose } from 'redux';
 import thunk from 'redux-thunk';
-import RootReducer from './reducers/RootReducer';
+import RootReducer from './reducers';
 
 const initialState = {};
 const middleWare = [thunk];
